+++ just_audio/CHANGELOG.md
@@ -1,5 +1,10 @@
 ## 0.11.0
 
+## 0.10.3
+
+* Fix pending timers bug in unit tests.
+* Fix NPE in Android position broadcast.
+
 ## 0.10.2
 
 * Add Player.playerEvent.